long type problem

+++ src/edu/kaist/mrlab/srdf/modules/DPWDChanger.java
@@ -72,7 +72,7 @@ public class DPWDChanger {
 				boolean stop = true;
 				while (dt.hasNext()) {
 					JSONObject innerDepen = (JSONObject) dt.next();
-					int depenID = (int) (long) innerDepen.get("id");
+					int depenID = (int) innerDepen.get("id");
 					String depenLabel = (String) innerDepen.get("label");
 					if (depenLabel.equals("NP_SBJ") && stop) {
 						firstSBJID = depenID;
@@ -97,15 +97,15 @@ public class DPWDChanger {
 				while (d.hasNext()) {
 
 					JSONObject innerDepen = (JSONObject) d.next();
-					int depenID = (int) (long) innerDepen.get("id");
-					int depenHead = (int) (long) innerDepen.get("head");
+					int depenID = (int) innerDepen.get("id");
+					int depenHead = (int) innerDepen.get("head");
 					String depenText = (String) innerDepen.get("text");
 					String depenLabel = (String) innerDepen.get("label");
 					JSONArray depenMod = (JSONArray) innerDepen.get("mod");
 					ArrayList<Integer> depenModArr = new ArrayList<Integer>();
 					Iterator<?> dm = depenMod.iterator();
 					while (dm.hasNext()) {
-						depenModArr.add((int) (long) dm.next());
+						depenModArr.add((int) dm.next());
 					}
 
 					// if(depenText.contains("한다는")){
@@ -129,9 +129,9 @@ public class DPWDChanger {
 
 						JSONObject innerWord = (JSONObject) w.next();
 						String wordText = (String) innerWord.get("text");
-						long wordID = (long) innerWord.get("id");
-						long wordBegin = (long) innerWord.get("begin");
-						long wordEnd = (long) innerWord.get("end");
+						int wordID = (int) innerWord.get("id");
+						int wordBegin = (int) innerWord.get("begin");
+						int wordEnd = (int) innerWord.get("end");
 
 						// System.out.println(wordID + "\t" + wordText + "\t"
 						// + wordBegin + "\t" + wordEnd);
@@ -147,7 +147,7 @@ public class DPWDChanger {
 						for (int k = 0; k <= wordEnd - wordBegin; k++) {
 
 							JSONObject innerMorp = (JSONObject) m.next();
-							long morpID = (long) innerMorp.get("id");
+							int morpID = (int) innerMorp.get("id");
 
 							String morpType = (String) innerMorp.get("type");
 							String morpLemma = (String) innerMorp.get("lemma");
@@ -193,21 +193,21 @@ public class DPWDChanger {
 										}
 
 										JSONObject VP2NP = (JSONObject) dma.next();
-										int dmaID = (int) (long) VP2NP.get("id");
+										int dmaID = (int) VP2NP.get("id");
 										String dmaLabel = (String) VP2NP.get("label");
 										if (dmaLabel.contains("VP")) {
 
 											int newHead = getNewHead(depenHead);
 											VP2NP.replace("head", newHead);
 											changeModOfNewHead(newHead, dmaID);
-											depenMod.remove((Object) (long) dmaID);
+											depenMod.remove((Object) (int) dmaID);
 
 										} else if (dmaLabel.contains("NP_SBJ") && dmaID == firstSBJID) {
 
 											int newHead = getNewHead(depenHead);
 											VP2NP.replace("head", newHead);
 											changeModOfNewHead(newHead, dmaID);
-											depenMod.remove((Object) (long) dmaID);
+											depenMod.remove((Object) (int) dmaID);
 
 										}
 									}
@@ -223,17 +223,17 @@ public class DPWDChanger {
 					lastDepen = (JSONObject) depenArr.get(depenArr.size() - 1);
 					JSONObject lastSecDepen = (JSONObject) depenArr.get(depenArr.size() - 2);
 
-					long lastDepenHead = (long) lastDepen.get("head");
+					int lastDepenHead = (int) lastDepen.get("head");
 					JSONArray lastDepenMod = (JSONArray) lastDepen.get("mod");
-					long lastDepenID = (long) lastDepen.get("id");
+					int lastDepenID = (int) lastDepen.get("id");
 					String lastDepenText = (String) lastDepen.get("text");
 					lastDepenLabel = (String) lastDepen.get("label");
 
 					// 뒤에서 두 번째 Word의 정보
 
-					long lastSecDepenHead = (long) lastSecDepen.get("head");
+					int lastSecDepenHead = (int) lastSecDepen.get("head");
 					JSONArray lastSecDepenMod = (JSONArray) lastSecDepen.get("mod");
-					long lastSecDepenID = (long) lastSecDepen.get("id");
+					int lastSecDepenID = (int) lastSecDepen.get("id");
 					String lastSecDepenText = (String) lastSecDepen.get("text");
 					String lastSecDepenLabel = (String) lastSecDepen.get("label");
 
@@ -266,7 +266,7 @@ public class DPWDChanger {
 	}
 
 	@SuppressWarnings("unchecked")
-	public void splitVNP(JSONObject targetDepen, long targetDepenHead, JSONArray targetDepenMod, long targetDepenID,
+	public void splitVNP(JSONObject targetDepen, int targetDepenHead, JSONArray targetDepenMod, int targetDepenID,
 			String targetDepenText, String targetDepenLabel, int index) {
 
 		boolean splitON = false;
@@ -274,9 +274,9 @@ public class DPWDChanger {
 		JSONObject targetWord = (JSONObject) wordArr.get(wordArr.size() - index);
 
 		String targetWordText = (String) targetWord.get("text");
-		long targetWordID = (long) targetWord.get("id");
-		long targetWordBegin = (long) targetWord.get("begin");
-		long targetWordEnd = (long) targetWord.get("end");
+		int targetWordID = (int) targetWord.get("id");
+		int targetWordBegin = (int) targetWord.get("begin");
+		int targetWordEnd = (int) targetWord.get("end");
 
 		Iterator<?> m = morpArr.iterator();
 
@@ -312,7 +312,7 @@ public class DPWDChanger {
 			}
 			for (int k = 0; k <= targetWordEnd - targetWordBegin; k++) {
 				JSONObject innerMorp = (JSONObject) m.next();
-//				long morpID = (long) innerMorp.get("id");
+//				int morpID = (int) innerMorp.get("id");
 				String morpType = (String) innerMorp.get("type");
 				String morpLemma = (String) innerMorp.get("lemma");
 
@@ -367,7 +367,7 @@ public class DPWDChanger {
 
 			while (dm.hasNext()) {
 				Iterator<?> da = depenArr.iterator();
-				long id = (long) dm.next();
+				int id = (int) dm.next();
 				for (int i = 0; i < id; i++) {
 					da.next();
 				}
@@ -384,8 +384,8 @@ public class DPWDChanger {
 			JSONArray newDepenMod = new JSONArray();
 
 			for (int i = 0; i < removeID.size(); i++) {
-				targetDepenMod.remove((long) removeID.get(i));
-				newDepenMod.add((long) removeID.get(i));
+				targetDepenMod.remove((int) removeID.get(i));
+				newDepenMod.add((int) removeID.get(i));
 			}
 
 			// System.out.println(NP + "\t" + VP);
@@ -468,7 +468,7 @@ public class DPWDChanger {
 		JSONObject tempDepen = (JSONObject) d.next();
 
 		int tempHead = Integer.parseInt(tempDepen.get("head").toString());
-		int tempID = (int) (long) tempDepen.get("id");
+		int tempID = (int) tempDepen.get("id");
 		String tempLabel = (String) tempDepen.get("label");
 //		String tempText = (String) tempDepen.get("text");
 
@@ -498,7 +498,7 @@ public class DPWDChanger {
 		}
 		JSONObject tempDepen = (JSONObject) d.next();
 		JSONArray tempMod = (JSONArray) tempDepen.get("mod");
-		tempMod.add((Object) (long) addMod);
+		tempMod.add((Object) (int) addMod);
 	}
 
 	public static void main(String[] ar) {
